Rename Score.h to game-source-code/Score.h
